LayoutTests/http/tests/media/video-referer-expected.txt

EVENT(canplay)
Tests that the media player will send the relevant referer when requesting the media file.
